// TicketReservation.UI/src/app/features/admin/events/models/event-model.ts
export interface EventModel {
  id: string;
  name: string;
  priceRangeEvent: string;
  priceRangeForTickets: string;
  description: string;
  eventDate: string;
  eventType: string;
  imageEvent: string;
  startTime: string;
  finishTime: string;
  locationId: string;
  locationName: string;
  seatMapPath: string;
}